Add files via upload
In this code, we define a custom annotation @MyAnnotation with a single attribute value. We then apply this annotation to a method myMethod in the MyClass class.

In the Main class, we demonstrate how to access and use the annotation. We retrieve the annotation from the myMethod using reflection and store it in a variable. Then, we access the value specified in the annotation and print it to the console.